internal/framework/flex: split options into standalone file
